refactor: :art: Create a `get_all_data` function and cache that
If we hashed each individual function in `transform`, which requires a
pd.DataFrame as input, we wouldnt be able to hash that. Instead, we
should wrap everything into a `get_all_data` function and only cache
that. We could pull out the DNO stuff as separate and cache to disk,
but it's also fine just to keep it a part of this function since we only
need it for intermediate transformations anyway, then we can discard it