Refactor `transform` and introduce a single `get_all_data` function (#16)
* refactor: :art: Rename `loaders` to `extract`

* refactor: :art: Move all data transformation outside of the `main` loop

In the next commit, we will refactor this code further

* refactor: :art: Create `get_regional_flex` function

Also refactor some of the other functions to reduce the amount of
intermediary dataframes we create

* refactor: :art: Make internal transform functions clearer

Prepend them with an underscore _

* perf(transform): :zap: Add caching to transform functions

Use `st.experimental_memo` to cache the output of functions.
We could add a `max_entries=1` in the future since that's all we
expect but I dont think we need it just now

* refactor: :art: Create a `get_all_data` function and cache that

If we hashed each individual function in `transform`, which requires a
pd.DataFrame as input, we wouldnt be able to hash that. Instead, we
should wrap everything into a `get_all_data` function and only cache
that. We could pull out the DNO stuff as separate and cache to disk,
but it's also fine just to keep it a part of this function since we only
need it for intermediate transformations anyway, then we can discard it